fix(lint): resolve all clang-tidy warnings in CI
- Remove set -e from lint-ci Makefile target; run-clang-tidy exits
  non-zero on any finding, causing the pipeline to abort before the
  warning/error counting logic could run
- Change highest_rssi type from int to int8_t to match ap_info[i].rssi
  (bugprone-signed-char-misuse)
- Replace nested ternary chains with if/else blocks in ui_binder_esp.c,
  loc_server.c, main.c and wifi_popup.c
  (readability-avoid-nested-conditional-operator)
- Add NOLINT for intentional int-to-ptr casts in LVGL user_data pattern
  (performance-no-int-to-ptr)
- Add NOLINT for oversized functions that cannot be meaningfully split:
  app_main, job_poll, handle_ws_message, build_password_popup

+++ components/wifi_manager/wifi_manager_esp.c
@@ -155,7 +155,7 @@ void wifi_manager_connect_to_saved_wifi(void) {
     esp_wifi_scan_get_ap_records(&ap_count, ap_info);
 
     int best_saved_index = -1;
-    int highest_rssi     = -100;
+    int8_t highest_rssi  = -100;
 
     for (int i = 0; i < ap_count; i++) {
         for (int j = 0; j < saved_count; j++) {
